requirements.txt
transformers
torch>=2.6.0
pillow
numpy<2.0
tensorflow
